# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: waitlist
    description: Waitlist and email OTP operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /send-otp:
    post:
      operationId: sendWaitlistOtp
      tags: [waitlist]
      summary: Send waitlist email OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendOtpResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /verify-otp:
    post:
      operationId: verifyWaitlistOtp
      tags: [waitlist]
      summary: Verify waitlist email OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpRequest"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyOtpResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /waitlist:
    post:
      operationId: createWaitlistUser
      tags: [waitlist]
      summary: Create a verified waitlist user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaitlistRequest"
      responses:
        "201":
          description: Waitlist user created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Email verification required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Duplicate waitlist registration
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SendOtpRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        fullName:
          type: string
          maxLength: 100
        mobile:
          type: string
          maxLength: 30
        city:
          type: string
          maxLength: 80
        message:
          type: string
          maxLength: 1000
      required:
        - email
    SendOtpResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
        expiresInSeconds:
          type: integer
      required:
        - success
        - message
        - expiresInSeconds
    VerifyOtpRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        otp:
          type: string
          pattern: "^\\d{6}$"
      required:
        - email
        - otp
    VerifyOtpResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
        expiresInSeconds:
          type: integer
      required:
        - success
        - message
        - expiresInSeconds
    WaitlistRequest:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
        mobile:
          type: string
          minLength: 7
          maxLength: 30
        city:
          type: string
          minLength: 2
          maxLength: 80
        message:
          type: string
          maxLength: 1000
      required:
        - fullName
        - email
        - mobile
        - city
    WaitlistResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
        id:
          type: string
          format: uuid
      required:
        - success
        - message
        - id
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          const: false
        error:
          type: string
      required:
        - success
        - error